+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Fixes
 * [#664](https://gitlab.com/meltano/meltano/issues/664) Minor CSS fix ensuring Airflow UI height is usable (side-effect of recent reparenting)
 * [#663](https://gitlab.com/meltano/meltano/issues/663) Fixed Airflow error when advancing to Orchestration step after installing and saving a Loader configuration
+* [#650](https://gitlab.com/meltano/meltano/issues/650) Add MELTANO_DISABLE_TRACKING ENV variable to disable all tracking
+* [#670](https://gitlab.com/meltano/meltano/issues/670) Update all Meltano tests to not send tracking Events
 
 ### Breaks